chore: 커뮤니티 관련 constants 파일 이동
